Moves the tests of jit.rs into separate files.
